util: Fix gethostname() and freeaddrinfo() calls
I was passing the wrong size to the 'gethostname()' call because the
'hostname' is now dynamically allocated and cannot use 'sizeof' to get
its size.  This mistake causes that the obtained hostname is empty.

The 'freeaddrinfo()' was placed at the wrong place.  Fix it.

Fallback to '_POSIX_HOST_NAME_MAX' if 'sysconf()' fails.

Always null-terminate the 'hostname'.