fixed gethostname() and freeaddrinfo() calls (#73)
* util: Fix gethostname() and freeaddrinfo() calls

I was passing the wrong size to the 'gethostname()' call because the
'hostname' is now dynamically allocated and cannot use 'sizeof' to get
its size.  This mistake causes that the obtained hostname is empty.

The 'freeaddrinfo()' was placed at the wrong place.  Fix it.

Fallback to '_POSIX_HOST_NAME_MAX' if 'sysconf()' fails.

Always null-terminate the 'hostname'.

* util: Simplify hostname() as no need to call getaddrinfo()

There is no need to call 'getaddrinfo()' to determine the hostname of
the machine, gethostname() is enough.  This simplify the 'hostname()'
a lot.

Trim a trailing whitespace by the way.