Fix ioctl issues on DragonFly BSD and FreeBSD (#78)
* Fix ioctl issues on DragonFly BSD and FreeBSD

* Include the correct headers on DragonFly BSD and FreeBSD.

* Use UTF-8 characters to draw border on Linux.

* Use ioctl 'KDGETLED' to get the keyboard LED states, which is supported
  on both Linux and BSD.

* Use macros 'LED_NUM' and 'LED_CAP' instead of hard-coded magic numbers.

NOTE:
This patch only allows 'ly' to build fine on DragonFly BSD and FreeBSD
(other *BSD not tested), and more works are needed to make 'ly' working
fine on BSD.

* Use ioctl 'KDGKBLED' on Linux

The legacy ioctl 'KDGETLED' (which also exists on BSD) is used to get the
state of keyboard LEDs, which, however, can be used to display arbitrary
information!  So the new ioctl 'KDGKBLED' should be used to get the keyboard
flags (CapsLock, NumLock, ScrollLock), and this ioctl has a separate set of
macros ('K_NUMLOCK', 'K_CAPSLOCK') to check the flags.

See the ioctl_console(2) man page for more details.