Unset DISPLAY variable for wayland
See #167 
This is the fix that I currently use to get ly to work with sway. I am not sure if this is the proper way to implement it as I have only tested it on bash and zsh. Also to note is that it does not fix the problem of XDG_SESSION_TYPE being set to `tty`