Set locale env variables in correct order
This commit aims to fix the use case of setting multiple environment
variables in locale settings. As described in manpages, consider the
following content of `/etc/locale.conf`:

    LANG=de_DE.UTF-8
    LC_MESSAGES=en_US.UTF-8

The previous behavior caused only `LANG` to be set, leaving the other
variables intact. I am not familiar how exactly shell interprets setting
values on these values, but the value of `LC_MESSAGES` in the example
was never applied to the environment. The only way how to make it
working was to execute `unset LANG` before the `locale.conf` variables
were set.

This code snippet is heavily inspired by
https://github.com/linuxdeepin/lightdm/blob/810f9b171feb779e48eba1dca38eb37d6d027d1a/src/session-child.c